chore: work on build warnings

+++ WoWMarketWatcher.API/Extensions/EntityFrameworkExtensions.cs
@@ -1,3 +1,4 @@
+using System;
 using Microsoft.EntityFrameworkCore;
 
 namespace WoWMarketWatcher.API.Extensions
@@ -6,6 +7,11 @@ namespace WoWMarketWatcher.API.Extensions
     {
         public static void Clear<T>(this DbSet<T> dbSet) where T : class
         {
+            if (dbSet == null)
+            {
+                throw new ArgumentNullException(nameof(dbSet));
+            }
+
             dbSet.RemoveRange(dbSet);
         }
     }
